+++ apps/server/src/modules/learnroom/repo/index.ts
@@ -3,10 +3,10 @@ export { ColumnBoardNode, ColumnBoardNodeProps } from './mikro-orm/column-board-
 export { ColumnBoardNodeRepo } from './mikro-orm/column-board-node.repo';
 export { CourseMikroOrmRepo } from './mikro-orm/course.repo';
 export { DashboardElementRepo } from './mikro-orm/dashboard-element.repo';
-export { DashboardModelMapper } from './mikro-orm/dashboard.model.mapper';
 export { DashboardRepo } from './mikro-orm/dashboard.repo';
 export { LegacyBoardElement, LegacyBoardElementType } from './mikro-orm/legacy-board-element.entity';
 export { BoardProps, LegacyBoard } from './mikro-orm/legacy-board.entity';
 export { LegacyBoardRepo } from './mikro-orm/legacy-board.repo';
 export { LessonBoardElement } from './mikro-orm/lesson-board-element.entity';
+export { DashboardModelMapper } from './mikro-orm/mapper/dashboard.model.mapper';
 export { TaskBoardElement } from './mikro-orm/task-board-element.entity';
